BAU: Rename index controller file
To be consistent with the others
